added folder structure and refactored folders
