--- sloby-editor-framework/framework_testing/index.ts
@@ -1,5 +1,0 @@
-function slobyTest(string: string) {
-  return string === "Sloby"
-}
-
-module.exports = slobyTest
