--- sloby-editor-framework/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "sloby-editor",
-  "version": "1.0.0",
-  "description": "This is the editor source code for the Sloby project",
-  "main": "./framework_testing/index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [
-    "Sloby"
-  ],
-  "author": "Hadhazy Gabor Co-founder of Sloby",
-  "license": "ISC"
-}
