+++ src/sloby-editor-framework/.gitignore
