+++ src/sloby-editor-framework/SlobyEditor/Editor.tsx
@@ -1,6 +1,6 @@
 import React from "react"
-import { AdminPageContainer } from "../../src/Editor/utils/styles/AdminPage"
-import { SlobyEditorContainer, SlobyConstantTools } from "../../src/Editor/utils/styles/Editor"
+import { AdminPageContainer } from "../../Editor/utils/styles/AdminPage"
+import { SlobyEditorContainer, SlobyConstantTools } from "../../Editor/utils/styles/Editor"
 import SlobyPreviewSite from "./SlobyPreviewSite"
 import SlobyTools from "./SlobyTools"
