+++ src/sloby-editor-framework/SlobyEditor/SlobyPreviewSite.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { SlobyPreviewSiteContainer } from '../../src/Editor/utils/styles/Editor'
+import { SlobyPreviewSiteContainer } from '../../Editor/utils/styles/Editor'
 
 function SlobyPreviewSite() {
   return <SlobyPreviewSiteContainer>
