+++ src/sloby-editor-framework/creators/detector/index.ts
