+++ src/sloby-editor-framework/framework_testing/index.ts
@@ -1,0 +1,1 @@
+import React from "react";
